Add Column non-expense release from restriction.
During Conservancy's FY 2015 audit, the auditors pointed out the
strangeness of this report summing 'Income' as negative.  Under the
current design of the restricted fund system using a duplicate chart of
accounts hierarchy for each temporary restricted asset, the logical way
to move funds between temp restricted funds and/or release them through
a mechanism other than an Expense is to have a transaction of this
nature:

  2015-12-24 Release Funds for general use per project instructions
        ;IncomeType: Donations
        ;Entity: Our-Organization
     Income:General:Donations:Released          $-1,000.00
     Income:Project:Donations:Released           $1,000.00

The problem with this sort of entry is that this report previously would
cause "Project" to have a debit in the Income column for this report.

This change specifically looks for 'Released' funds and puts them in a
different column.