Fix tag value parsing
If a tag is more than 2 characters from the beginning of the comment the
tag value offset will be wrong. #1702 gives an example where the tag
line starts with `;;` and the tag value thus becomes `: Bar` because of
this bug.

The use `index` in the offset calulation seems to be a lucky coincidence
that works in the common case: "; tag: value"

Fixes #1702