Increase all buffers 8k⇒64k since error output limit is so increased
In an abundance of caution, to match the increase in error output limit
provided by John Wiegley <johnw@newartisans.com>, increase other buffer sizes
that were limited to 8k throughout the codebase to 64k.

This may not be necessary and indeed I was a bit worried to find so many hard
limits across the code, but felt that updating the all the buffers to match
that number made sense.

It would be useful if someone went through and figured out whether these
buffer limits should be there and if they impact usage in some way.